feat: runner and the first test
