+++ Cargo.toml
@@ -8,6 +8,7 @@ edition = "2021"
 [dependencies]
 anyhow = "1.0.66"
 once_cell = "1.16.0"
+pretty_assertions = "1.3.0"
 regex = "1.7.0"
 thiserror = "1.0.37"
 wasmer = "3.0.2"
